MOTION_DETECTOR
The disability targetted here is ADHD (Attention Deficit Hyperactivity Disorder). It is difficult for students with ADHD to study effectively and the goal of this software is to help such students. (Anyone with a low attention span can use the software)

The software constantly checks the status of student and gives stress relieving “breathe in and breathe out prompts” along with a final analysis which can be used for further tracking.